backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.3.1
django-filter==23.4
python-decouple==3.8
psycopg2-binary==2.9.9
Pillow==10.1.0
drf-spectacular==0.26.5
gunicorn==21.2.0
